fix long id, added occupant relation

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/daos/SimpleDao.java
@@ -5,6 +5,6 @@ import java.util.Optional;
 
 public interface SimpleDao<T> {
     T save(T t);
-    Optional<T> findById(int id);
+    Optional<T> findById(long id);
     List<T> findAll();
 }
